Deal with the vulnerability that a malicious user could place wastebin emojis on all Codeys messages, including ones that may be important, or ones not intended to be deleted.

+++ src/main/java/io/horrorshow/codey/discordutil/CommonJDAListener.java
@@ -8,6 +8,8 @@ import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
 
@@ -22,6 +24,8 @@ public class CommonJDAListener extends ListenerAdapter {
         jda.addEventListener(this);
     }
 
+    //Being static avoids the need to inject this map to DiscordUtils
+    public static Map<String,Boolean> deletableMessageMap = new HashMap<>();
 
     @Override
     public void onGuildMessageReactionAdd(@NotNull GuildMessageReactionAddEvent event) {
@@ -33,9 +37,19 @@ public class CommonJDAListener extends ListenerAdapter {
 
     public void onReactionAdd(@NotNull GuildMessageReactionAddEvent event) {
         var message = event.getChannel().retrieveMessageById(event.getMessageId()).complete();
+        /***
+         * Use a hashmap that is not persisted in any way to look up if the message should be deletable.
+         *
+         * It should not be a big issue, if a user reacts, and the message is not deleted.
+         * However, vice-versa is not true. If a malicious user reacts, and the message is deleted - that is an issue.
+         *
+         * The fact that the hashmap is empty on restart does not matter too much.
+         */
+
         if (DiscordUtils.hasEmoji(BASKET, event)) {
-            if (event.getJDA().getSelfUser().getId().equals(message.getAuthor().getId())) {
+            if (deletableMessageMap.containsKey(event.getMessageId()) && event.getJDA().getSelfUser().getId().equals(message.getAuthor().getId())) {
                 try {
+                    deletableMessageMap.remove(message.getId());
                     message.delete().complete();
                 } catch (ErrorResponseException e) {
                     log.warn("Unable to remove message");
